src/requirements.txt
pillow>=10.3.0
pyreadline3>=3.5.4; sys_platform=='win32'
fontTools>=4.51.0